+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 
 All notable changes to `file-url-signer` will be documented in this file
 
+# 1.0.3
+
+- Integer less than 10 bug to hexadecimal fixed;
+- Added extra query parameters to `validate()` method output.
+
 # 1.0.2
 
 - File Entity class scan and extract parameters from the file name;